+++ example/lib/pages/dropdown_field_page.dart
@@ -75,7 +75,7 @@ class DropdownFieldPage extends StatelessWidget {
                 items: <String>[
                   'Lorem ipsum dolor sit amet',
                   'consectetur adipiscing elit',
-                  'sed do eiusmod tempor incididunt ut labore et dolore magna aliqua',
+                  'sed do eiusmod tempor incididunt ut labore et dolore magna',
                 ],
                 onChanged: print,
                 prefixIcon: Icon(Icons.density_large),