+++ lib/src/date_range_field.dart
@@ -29,8 +29,6 @@ class DateRangeField extends StatefulWidget {
   final String? errorText;
 
   /// The first date of the date picker.
-  ///
-  /// This defaults to [DateTime.now].
   final DateTime? firstDate;
 
   /// Determines if a error should be shown.
@@ -40,8 +38,6 @@ class DateRangeField extends StatefulWidget {
   final String? labelText;
 
   /// The last date of the date picker.
-  ///
-  /// This defaults to [DateTime.now.add(Duration(days: 365))].
   final DateTime? lastDate;
 
   /// An icon that appears before the editable part of the text field.
@@ -62,14 +58,19 @@ class _DateRangeFieldState extends State<DateRangeField> {
   late FocusNode _focusNode;
   DateTimeRange? _value;
 
-  DateTime get _defaultFirstDate => DateTime.now();
-  DateTime get _defaultLastDate =>
-      DateTime.now().add(const Duration(days: 365));
+  DateTime get _defaultFirstDate => DateTime.fromMillisecondsSinceEpoch(0);
+  DateTime get _defaultLastDate => DateTime(2100);
 
   @override
   void initState() {
     super.initState();
     _focusNode = FocusNode();
+    _value = widget.value;
+  }
+
+  @override
+  void didUpdateWidget(DateRangeField oldWidget) {
+    super.didUpdateWidget(oldWidget);
     _value = widget.value;
   }