+++ lib/src/save_button.dart
@@ -34,11 +34,9 @@ class SaveButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    // TODO: latest release flags this with "touch target size" warning
-    // this means the ElevatedButton needs to be bigger...
     return SizedBox(
       // this is to avoid a height change during the transition
-      height: Theme.of(context).buttonTheme.height,
+      height: _kDefaultButtonHeight,
       width: double.infinity,
       child: AnimatedSwitcher(
         duration: duration,